Move utility functions back to utils
